shopflow/backend/user-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
